refactor(dao): GameDaoImpl to GameDao
- delete static factory method ofProd and table field in GameDao
- remove GameDao interface
- create test stubs as children of GameDao
