+++ src/test/java/chess/dao/GameDaoTest.java
@@ -15,7 +15,12 @@ class GameDaoTest {
 
     private static final String CLEANSE_TEST_DB_SQL = String.format("TRUNCATE TABLE %s", TEST_TABLE);
 
-    private final GameDao dao = new GameDaoImpl(TEST_TABLE);
+    private final GameDao dao = new GameDao() {
+        @Override
+        protected String addTable(String sql) {
+            return String.format(sql, TEST_TABLE);
+        }
+    };
 
     @BeforeEach
     void setUp() {
